+++ NEWS.md
@@ -1,10 +1,15 @@
-# AMR 3.0.1.9006
+# AMR 3.0.1.9007
 
 ### New
 * Integration with the **tidymodels** framework to allow seamless use of SIR, MIC and disk data in modelling pipelines via `recipes`
   - `step_mic_log2()` to transform `<mic>` columns with log2, and `step_sir_numeric()` to convert `<sir>` columns to numeric
-  - New `tidyselect` helpers: `all_sir()`, `all_sir_predictors()`, `all_mic()`, `all_mic_predictors()`, `all_disk()`, `all_disk_predictors()`
+  - New `tidyselect` helpers:
+    - `all_sir()`, `all_sir_predictors()`
+    - `all_mic()`, `all_mic_predictors()`
+    - `all_disk()`, `all_disk_predictors()`
 * Data set `esbl_isolates` to practise with AMR modelling
+* AMR selectors `phosphonics()` and `spiropyrimidinetriones()`
+* `ab_group()` gained an argument `all_groups` to return all groups the antimicrobial drug is in (#246)
 
 ### Changed
 * Fixed a bug in `antibiogram()` for when no antimicrobials are set
@@ -16,7 +21,9 @@
   * Interpretation of capped MIC values now consistently returns `"NI"` (non-interpretable) when the true MIC could be at either side of a breakpoint, depending on the selected handling mode
   * This results in more reliable behaviour compared to previous versions for capped MIC values
   * Removed the `"inverse"` option, which has now become redundant
-
+* Fixed some foreign translations of antimicrobial drugs
+* `antimicrobials$group` is now a `list` instead of a `character`, to contain any group the drug is in (#246)
+* `ab_group()` now returns values consist with the AMR selectors (#246)
 
 # AMR 3.0.1